crypto: drop support for Botan and OpenSSL variants
For full discussion please refer to #3391 .
