+++ src/plugins/crypto/README.md
@@ -21,32 +21,11 @@ It also needs to be decrypted whenever an admissible access (read) is being perf
 The users of Elektra should not be bothered too much with the internals of the cryptographic operations.
 Also the cryptographic keys must never be exposed to the outside of the crypto module.
 
-The crypto plugin supports different libraries as provider for the cryptographic operations.
-At the moment the following crypto APIs are supported:
-
-- OpenSSL (`libcrypto`)
-- libgcrypt
-- Botan
+The crypto plugin uses libgcrypt as provider of cryptographic operations.
 
 ## Dependencies
 
-#ifdef ELEKTRA_CRYPTO_API_GCRYPT
-
 - `libgcrypt20-dev` or `libgcrypt-devel`
-
-#endif
-
-#ifdef ELEKTRA_CRYPTO_API_OPENSSL
-
-- `libssl-dev` or `openssl-devel`
-
-#endif
-
-#ifdef ELEKTRA_CRYPTO_API_BOTAN
-
-- `libbotan1.10-dev` or `botan-devel`
-
-#endif
 
 ### GnuPG (GPG)
 
@@ -58,39 +37,23 @@ If no such configuration is provided, the plugin will look at the PATH environme
 
 ## How to compile
 
-The following compilation variants are available:
-
-1. crypto_gcrypt
-2. crypto_openssl
-3. crypto_botan
-
-Add "crypto" and the variants, that you want (you can add one of them or all), to the `PLUGINS` variable in `CMakeCache.txt` and re-run `cmake`.
-
-In order to add all compile variants you can add "CRYPTO" to the `PLUGINS` variable.
+Add "crypto" to the `PLUGINS` variable in `CMakeCache.txt` and re-run `cmake`.
 
 An example `CMakeCache.txt` may contain the following variable:
 
 ```cmake
-PLUGINS=crypto;crypto_gcrypt;crypto_openssl;crypto_botan
-```
-
-or it may look like:
-
-```cmake
-PLUGINS=CRYPTO
+PLUGINS=crypto
 ```
 
 ### macOS
 
-All variants of the plugin work under macOS Sierra (Version 10.12.3 (16D32)).
+The crypto plugin works under macOS Sierra (Version 10.12.3 (16D32)).
 
 To set up the build environment on macOS Sierra we recommend using [Homebrew](http://brew.sh/).
 Follow these steps to get everything up and running:
 
 ```sh
-brew install openssl botan libgcrypt pkg-config cmake
-# The next step is required for pkg-config to find the include files of OpenSSL
-ln -s /usr/local/opt/openssl/include/openssl/ /usr/local/include/openssl
+brew install libgcrypt pkg-config cmake
 ```
 
 Also a GPG installation is required. The [GPG Tools](https://gpgtools.org) work fine for us.
@@ -101,10 +64,10 @@ At the moment the plugin will only run on Unix/Linux-like systems, that provide 
 
 ## Examples
 
-To mount a backend with the gcrypt plugin variant that uses the GPG key 9CCC3B514E196C6308CCD230666260C14A525406, use:
+To mount a backend with the crypto plugin that uses the GPG key 9CCC3B514E196C6308CCD230666260C14A525406, use:
 
 ```sh
-kdb mount test.ecf user/t crypto_gcrypt "crypto/key=9CCC3B514E196C6308CCD230666260C14A525406"
+kdb mount test.ecf user/t crypto "crypto/key=9CCC3B514E196C6308CCD230666260C14A525406"
 ```
 
 Now you can specify a key `user/t/a` and protect its content by using:
@@ -148,7 +111,7 @@ If you are not sure which keys are available to you, the `kdb` program will give
 For example you can type:
 
 ```sh
-kdb mount test.ecf user/t crypto_gcrypt
+kdb mount test.ecf user/t crypto
 ```
 
 In the error description you should see something like:
@@ -170,7 +133,7 @@ This means that the following keys are available:
 So the full mount command could look like this:
 
 ```sh
-kdb mount test.ecf user/t crypto_gcrypt "crypto/key=847378ABCF0A552B48082A80C52E8E92F785163F"
+kdb mount test.ecf user/t crypto "crypto/key=847378ABCF0A552B48082A80C52E8E92F785163F"
 ```
 
 ### Cryptographic Operations
@@ -206,6 +169,4 @@ Shutdown is enabled in the unit tests to prevent memory leaks.
 
 ### Ciphers and Mode of Operation
 
-All of the plugin variants use the Advanced Encryption Standard (AES) in Cipher Block Chaining Mode (CBC) with a key size of 256 bit.
-
-The ciphers and modes of operations are defined in the corresponding `<plugin_variant>_operations.c` or `<plugin_variant>_operations.h` file.
+The crypto plugin uses the Advanced Encryption Standard (AES) in Cipher Block Chaining Mode (CBC) with a key size of 256 bit.
